pyproject.toml
[project]
name = "zorvyn-assignment"
version = "0.1.0"
description = ""
authors = [
    {name = "Biswajit Das",email = "mrbiswajitdas3@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard] (>=0.135.3,<0.136.0)",
    "tortoise-orm (>=1.1.7,<2.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)"
]

[dependency-groups]
dev = [
    "python-dotenv (>=1.2.2,<2.0.0)",
    "ruff (>=0.15.9,<0.16.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
